% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aframer.R
\name{a_map}
\alias{a_map}
\title{Aframer Map}
\usage{
a_map(
  lat = 45.9,
  lon = 6.8,
  radius = 5,
  zoom = 12,
  axes_on = FALSE,
  mapbox_token = Sys.getenv("MAPBOX_TOKEN"),
  rotation = aframer::xyz_aframe(-90, 180, 0),
  position = aframer::xyz_aframe(0, 0.7, -1.5),
  scale = aframer::xyz_aframe(2, 2, 2),
  ...
)
}
\arguments{
\item{lat}{Center coordinates of map.}

\item{lon}{Center coordinates of map.}

\item{radius}{The radius of the circle that fits the terrain.}

\item{zoom}{Satellite zoom resolution of the tiles in the terrain. 
Select from {11, 12, 13, 14, 15, 16, 17}, where 17 is the highest 
value supported. For a fixed radius, higher zoom resolution results 
in more tileset API calls.}

\item{axes_on}{Whether to draw axes.}

\item{mapbox_token}{\href{https://www.mapbox.com/}{Mapbox} API token.}

\item{rotation}{Rotation of Map, default should work.}

\item{position}{Position of map on scene.}

\item{scale}{Scale of map.}

\item{...}{Any other options passed to the aframe entity.}
}
\description{
Create Virtual Reality Map.
}
