% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{everest}
\alias{everest}
\title{Everest}
\format{An data.frame of 5 rows.
\describe{
  \item{\code{name}}{ Camp name}
  \item{\code{lat}}{ Latitude}
  \item{\code{lon}}{ Longitude}
  \item{\code{ele}}{ Elevation in meters}
}}
\usage{
everest
}
\description{
Manually collected coordinates on Mount Everest base camps and summit: 
don't rely on this for your next hike.
}
\keyword{datasets}
