% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aframer.R
\name{g3r_aframer_dependency}
\alias{g3r_aframer_dependency}
\title{Dependency}
\usage{
g3r_aframer_dependency()
}
\description{
Dependency for g3r aframe component.
}
\details{
Include in your \link[aframer]{a_scene}.
}
