% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{g3r_camera}
\alias{g3r_camera}
\title{Adjust Camera Position}
\usage{
g3r_camera(g3r, lat = 0, lon = 0, elevation = 1.5)
}
\arguments{
\item{g3r}{An object of class \code{g3r} as returned by \code{g3r}.}

\item{lat, lon, elevation}{Position of camera.}
}
\description{
Adjust the position of the camera.
}
