% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{g3r_demo}
\alias{g3r_demo}
\title{Demo Application}
\usage{
g3r_demo()
}
\description{
Demo Application that displays the 3D map of your surroundings.
}
\details{
Requires the \href{https://github.com/ColinFay/geoloc}{geoloc}
package installed.
}
\examples{
\dontrun{g3r_demo}

}
