% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{g3r_lines}
\alias{g3r_lines}
\title{Lines}
\usage{
g3r_lines(g3r, data, lat, lon, elevation, width = 2L, color = "blue")
}
\arguments{
\item{g3r}{An object of class \code{g3r} as returned by \code{g3r}.}

\item{data}{A data.frame holding coordinates to plot.}

\item{lat}{Position of camera.}

\item{lon}{Position of camera.}

\item{elevation}{Position of camera.}

\item{width}{The width of the line.}

\item{color}{Color of the point.}
}
\description{
Add lines to a map.
}
\examples{
\dontrun{ 
g3r(27.985655, 86.903697) \%>\% 
  g3r_lines(nuptse, lat, lon, ele) 
}

}
