% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{g3r_points}
\alias{g3r_points}
\title{Points}
\usage{
g3r_points(g3r, data, lat, lon, elevation, size, color)
}
\arguments{
\item{g3r}{An object of class \code{g3r} as returned by \code{g3r}.}

\item{data}{A data.frame holding coordinates to plot.}

\item{lat, lon}{Coordinates and elevation of point.}

\item{elevation}{Elevation of point in meters.}

\item{size}{The size of the point.}

\item{color}{Color of the point.}
}
\description{
Add points to a map.
}
\examples{
\dontrun{ 
g3r(46.5763, 7.9904) \%>\% 
  g3r_points(everest, lat, lon, ele) 
}

}
