% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{get_elevations}
\alias{get_elevations}
\title{Get Elevations of point}
\usage{
get_elevations(
  lat,
  lon,
  limit = 50L,
  mapbox_token = Sys.getenv("MAPBOX_TOKEN")
)
}
\arguments{
\item{lat}{Center coordinates of map.}

\item{lon}{Center coordinates of map.}

\item{limit}{Maximum number of elevations to return.}

\item{mapbox_token}{\href{https://www.mapbox.com/}{Mapbox} API token.}
}
\value{
A vector of elevations.
}
\description{
Get elevations of given coordinates using the Mapbox tilequery API.
}
\examples{
\dontrun{get_elevations(39.75953, -105.01109)}

}
