% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{g3r_points}
\alias{g3r_points}
\title{Points}
\usage{
g3r_points(g3r, data, lat, lon, elevation, color)
}
\arguments{
\item{g3r}{An object of class \code{g3r} as returned by \code{g3r}.}

\item{data}{A data.frame holding coordinates to plot.}

\item{lat, lon}{Coordinates and elevation of point.}

\item{elevation}{Elevation of point in meters.}

\item{color}{Color of the point.}
}
\description{
Add points to a map.
}
\examples{
\dontrun{
df <- data.frame(
  lon =  7.9610, 
  lat = 46.5852,
  ele = 2100
)

g3r(46.5763, 7.9904, width = 900) \%>\% 
  g3r_points(df, lat, lon, ele) 
}

}
